decoder_service/requirements.txt
fastapi==0.115.12
pillow==11.2.1
pyzbar==0.1.9
python-multipart==0.0.20
uvicorn==0.34.0
